Create first version of CV
Ok, there is some hackery in here. Currently the vitae::detailed_entries function cannot handle markdown styled text which is annoying because all of my data on the google sheets includes this. So I had to create a bunch of functions to clean stuff out and reconfigure some other things. The results is a decent looking CV in PDF format.